parse PRIVATE key in binding file
When a binding does not contain a value for the PRIVATE key, then DHCP
responses contain a default gateway, otherwise they don't. This is
useful when a VM has multiple DHCP enabled interfaces to avoid getting
into the situation where a private network's gateway overrides the
gateway of the public network.