Implement IPv6 mode configuration parameter
This parameter can be used to control whether SLAAC or SLAAC with
stateless DHCPv6 will be used on IPv6 setups. Those can be configured
explicitly by providing `slaac' or `slaac+dhcpv6' as value for this new
parameter, or use `auto' to allow nfdhcpd to determine itself the
running mode by checking the provided values for the ns, rs and dhcpv6
queues. If the value for the dhcpv6_queue is not defined, then SLAAC is
used.

Remove obsolete enable_dhcpv6 parameter.